# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestão Bancária",
        "version": "1.0.0"
    },
    "paths": {
        "/conta": {
            "get": {
                "summary": "Endpoint para consultar contas.",
                "description": "Retorna informações da conta específicada pelo parâmetro \n        \"id\" na query string ou, se o ID for omitido, retorna uma listagem de \n        todas as contas cadastradas",
                "operationId": "480e5be39c15bbb4621aeeec6d84a985",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "query",
                        "description": "ID da conta",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações de determinada conta ou uma listagem de todas contas ativas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContaResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Foi requisitada uma determinada conta, mas ela não existe ou está inativa."
                    }
                }
            },
            "post": {
                "summary": "Armazena a nova conta no banco de dados.",
                "description": "Cadastra uma nova conta para transações futuras.",
                "operationId": "420c9cfc292b183cf2260e3472afed2f",
                "requestBody": {
                    "description": "Informações sobre a conta no formato JSON. 'conta_id' \n            é opcional e, se for omitido, será geradado um ID automático. 'valor' \n            é o saldo inicial da conta.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "conta_id": {
                                        "title": "conta_id",
                                        "description": "ID da conta (opcional)",
                                        "type": "integer"
                                    },
                                    "valor": {
                                        "title": "valor",
                                        "description": "Saldo inicial da conta.",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Conta criada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContaResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Informações inválidas da conta."
                    },
                    "409": {
                        "description": "Já existe uma conta cadastrada com o mesmo ID."
                    }
                }
            }
        },
        "/conta/{id}": {
            "delete": {
                "summary": "Desativa determinada conta pelo ID.",
                "operationId": "0ed503be786e053255203c165e44973b",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "ID da conta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conta desativada com sucesso."
                    },
                    "404": {
                        "description": "Foi requisitada uma determinada conta, mas ela não existe ou ja estava inativa."
                    }
                }
            }
        },
        "/transacao": {
            "post": {
                "summary": "Realiza um transação entre duas contas bancárias.",
                "description": "Realizar a movimentação de saldos: subtrai do pagador \n        (junto com a taxa) e adiciona ao receber (sem taxa).",
                "operationId": "fda6c4053a2c26bf80a1226ae340e8ac",
                "requestBody": {
                    "description": "Informações necessárias para realizar a movimentação de \n            saldo entre as contas.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransacaoBancariaResource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transação realizada com sucesso. Retorna os dados \n                atualizados da conta do pagador.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContaResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Informações inválidas da transação."
                    },
                    "404": {
                        "description": "A conta do recebedor e/ou do pagador não foi encontrada\n                ou o saldo do pagador é insuficiente para realizar a transação."
                    }
                }
            }
        },
        "/transacao/{contaId}": {
            "get": {
                "summary": "Lista todas as transações de determinada conta.",
                "description": "Retorna uma lista com todas as transações de determinada conta.",
                "operationId": "35d1bab894d14e5143f4e50bba991911",
                "parameters": [
                    {
                        "parameter": "contaId",
                        "name": "contaId",
                        "in": "path",
                        "description": "ID da conta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uma listagem de todas contas ativas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransacaoBancariaResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "A conta requisitada não existe ou foi desativada."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ContaResource": {
                "title": "Conta",
                "required": [
                    "saldo"
                ],
                "properties": {
                    "conta_id": {
                        "title": "conta_id",
                        "type": "integer"
                    },
                    "saldo": {
                        "title": "saldo",
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "TransacaoBancariaResource": {
                "title": "Transação Bancária",
                "required": [
                    "recebedor_id",
                    "pagador_id",
                    "forma_pagamento",
                    "valor"
                ],
                "properties": {
                    "recebedor_id": {
                        "title": "recebedor_id",
                        "type": "integer"
                    },
                    "pagador_id": {
                        "title": "pagador_id",
                        "type": "integer"
                    },
                    "forma_pagamento": {
                        "title": "forma_pagamento",
                        "description": "Formas de pagamento aceitas: 'C' (Crédito), 'D' (Débito) e 'P' (Pix)",
                        "type": "string"
                    },
                    "valor": {
                        "title": "valor",
                        "type": "number"
                    }
                },
                "type": "object"
            }
        }
    }
}